+++ src/Client/ClientSetup.h
@@ -525,7 +525,6 @@ public:
             "no-wave",
             "no-particles",
             "solid-wave",
-            "no-shake",
             "custom-trans",
             "menu-bg-blur",
             "free-scroll",
@@ -574,10 +573,6 @@ public:
         #endif
 
         #ifdef GEODE_IS_WINDOWS
-        //Client::GetModule("tps-bypass")->setIncompatible("This mod has <cr>not yet</c> been ported to <cl>Windows 2.206</c>");
-        //Client::GetModule("all-plat")->setIncompatible("This mod has <cr>not yet</c> been ported to <cl>2.206</c> on <cl>Windows</c>.");
-        Client::GetModule("no-shake")->setIncompatible("This mod has <cr>not yet</c> been ported to <cl>2.206</c> on <cl>Windows</c>.");
-        //Client::GetModule("custom-obj-limit")->setIncompatible("This mod has <cr>not yet</c> been ported to <cl>2.206</c> on <cl>Windows</c>.");
 
         #endif